+++ src/rename-angular-component/in-file-edits/get-core-file-path.function.ts
@@ -2,6 +2,5 @@ import { FileToMove } from '../file-manipulation/files-related-to-stub.class';
 
 export function getCoreFilePath(filesToMove: FileToMove[]) {
   let coreConstructs = filesToMove.filter((f) => f.isCoreConstruct);
-  // TODO: add logic for handling core router modules
   return coreConstructs[0]?.filePath;
 }